add first reputation event enabled, increment reputation of farmer when they vote with consensus on a transaction
