+++ crates/events/Cargo.toml
@@ -35,6 +35,7 @@ cuckoofilter = { workspace = true }
 ethereum-types = { workspace = true }
 quorum = { workspace = true }
 messr = { workspace = true }
+decentrust = { workspace = true }
 
 [dev-dependencies]
 serial_test = { workspace = true }
